+++ app/javascript/vue/components/georeferences/map.vue
@@ -98,13 +98,17 @@ const props = defineProps({
     type: Boolean,
     default: true
   },
+  rotateMode: {
+    type: Boolean,
+    default: false
+  },
   tilesSelection: {
     type: Boolean,
     default: true
   },
   tooltips: {
     type: Boolean,
-    default: true
+    default: false
   },
   center: {
     type: Array,
@@ -129,6 +133,10 @@ const props = defineProps({
   zoomBounds: {
     type: Number,
     default: undefined
+  },
+  actions: {
+    type: Boolean,
+    default: false
   }
 })
 
@@ -184,6 +192,8 @@ onMounted(() => {
   addDrawControllers()
   handleEvents()
 
+  mapObject.pm.setGlobalOptions({ tooltips: props.tooltips })
+
   if (props.geojson.length) {
     geoJSON(props.geojson)
   }
@@ -236,7 +246,14 @@ const addDrawControllers = () => {
       editMode: props.editMode,
       dragMode: props.dragMode,
       cutPolygon: props.cutPolygon,
-      removalMode: props.removalMode
+      removalMode: props.removalMode,
+      rotateMode: props.rotateMode
+    })
+  }
+
+  if (!props.actions) {
+    mapObject.pm.Toolbar.getControlOrder().forEach(control => {
+      mapObject.pm.Toolbar.changeActionsOfControl(control, [])
     })
   }
 }